pyproject.toml
[project]
name = "discord-sentiment-bot"
version = "1.0.0"
description = "Discord bot for sentiment analysis using Claude API"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your@email.com"}
]

dependencies = [
    "nextcord>=2.6.0",
    "anthropic>=0.28.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
    "aiosqlite>=0.19.0",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
]

[project.scripts]
sentiment-bot = "bot.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
